Extract ObjectRelationships.styled
