+++ frontend/src/metabase/visualizations/components/ObjectDetail/ObjectDetailView.styled.tsx
@@ -1,7 +1,5 @@
 import styled from "@emotion/styled";
 
-import { color } from "metabase/lib/colors";
-
 interface ObjectDetailContainerProps {
   wide: boolean;
 }
@@ -20,13 +18,6 @@ export const ObjectDetailsTable = styled.div`
   overflow-y: auto;
   flex: 1;
   padding: 2rem;
-`;
-
-export const ObjectRelationships = styled.div`
-  overflow-y: auto;
-  flex: 0 0 100%;
-  padding: 2rem;
-  background-color: ${color("bg-light")};
 `;
 
 export const ErrorWrapper = styled.div`
@@ -58,22 +49,4 @@ export const FitImage = styled.img`
   max-height: 18rem;
   object-fit: contain;
   margin: 1rem auto;
-`;
-
-export interface ObjectRelationshipContentProps {
-  isClickable: boolean;
-}
-
-export const ObjectRelationContent = styled.div<ObjectRelationshipContentProps>`
-  display: flex;
-  align-items: center;
-  margin: 1rem 0;
-  padding-bottom: 1rem;
-  border-bottom: 1px solid ${color("border")};
-  color: ${props => color(props.isClickable ? "text-dark" : "text-medium")};
-  cursor: ${props => props.isClickable && "pointer"};
-
-  &:hover {
-    color: ${props => props.isClickable && color("brand")};
-  }
 `;
